payment method session
